Initial commit, no GMP, naive algorithms
This is the initial commit, the first working version. It is monolithic,
with mostly naive operations (especially polynomial multiplication), and
not using GMP. The next tasks are probably to separate the different
kinds of functions into modules (polynomial functions, modular
arithmetic functions, and the main AKS function), switch to GMP for
arbitrary precision, implement FFT for polynomial multiplcation,
implement Euler's totient and use it for finding a better bound on the
main loop of step 5 of AKS, prepare a 'database' of test cases to test
against, take emprical time measurements and/or do more extensive
profiling.